requirements.txt
discord.py
python-dotenv
langchain-core
langchain-openai
langsmith
